Refs codership/galera#275, codership/galera#274 - this commit fixes 32-bit compilation issue in previous commit and also makes further cleanups with respect to type safety. In particular it does away with use of ssize_t, making all store and configuration sizes size_t.

+++ gcache/src/gcache_rb_store.hpp
@@ -91,7 +91,7 @@ namespace gcache
             if (next_ >= first_)
                 assert(0 == size_trail_);
             else
-                assert(size_trail_ >= ssize_t(sizeof(BufferHeader)));
+                assert(size_trail_ >= sizeof(BufferHeader));
 #endif
         }
